added caching.
In general getting content by filepath's was sending 2 requests,
which reaches the request-limit of 60 very fast. Even though the
limits are on an IP basis this caching scheme seems best.

I also realize that ajax calls have a `cache: true`. However, this
was not working for `$.getGithubFile` for some reason, and this
method works absolutely.